OTFeatureInfo: add some general to display what the features do
NOTE: This is only limited cool. Some fonts don't even implement
the minimum one would expect. At least we get to see roughly what
we are up to.

To make this widget more useful, we could think about ways to
configure the example texts per specimen. One question is if
we'd need a per font config option or if a config option for all
fonts in a specimen would suffice.