Initial commit Diagram widget
The aim of the Diagram widget is to enable the rendering of SVG
based diagrams to illustrate certain aspects of the font.
Examples are right now a rather generic x-height diagram and a
rather specific `stylisticSets` diagram, that renders two glyph
alternates on top of each other.
Eventually, a small domain specific language or something similar
should be available to describe such diagrams, so that it is easy
for a designer or a specialized tool, to create these diagrams
for specific fonts.